Refatoração de Double para BigDecimal, com testes.

+++ testes/br/loja/dominio/TestesProduto.java
@@ -1,5 +1,7 @@
 package br.loja.dominio;
 
+import java.math.BigDecimal;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -12,19 +14,19 @@ public class TestesProduto {
 	@Test
 	public void naoDeveriaAceitarProdutosComPrecoNegativo() {
 		thrown.expect(IllegalArgumentException.class);
-		new Produto("PRD-0009", -559.99);
+		new Produto("PRD-0009", BigDecimal.valueOf(-559.99));
 	}
 
 	@Test
 	public void deveriaCriarProduto() {
-		new Produto("PRD-0009", 559.99);
+		new Produto("PRD-0009", BigDecimal.valueOf(559.99));
 	}
 
 	@Test
 	public void naoDeveriaAlterarPrecoParaNegativo() {
 		thrown.expect(IllegalArgumentException.class);
-		Produto iPhone6 = new Produto("PRD-0001", 2399.99);
-		iPhone6.setPreco(Double.valueOf(-1));
+		Produto iPhone6 = new Produto("PRD-0001", BigDecimal.valueOf(2399.99));
+		iPhone6.setPreco(BigDecimal.valueOf(-1));
 	}
 
 }
